requirements.txt
pyModbusTCP>=0.2.0
numpy>=1.24.3,<2.0.0
pyyaml>=6.0.1,<7.0.0
fastapi>=0.110.0
uvicorn>=0.27.1
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.6.3
paho-mqtt>=1.6.1 